refactor(car enums): conform with change at car enums
